fixes #4.

+++ src/main/java/sds/classfile/attributes/annotation/ParameterAnnotations.java
@@ -15,7 +15,7 @@ public class ParameterAnnotations {
 	private Annotation[] annotations;
 
 	ParameterAnnotations(RandomAccessFile raf) throws IOException, ElementValueException {
-		this.annotations = new Annotation[raf.readByte()];
+		this.annotations = new Annotation[raf.readShort()];
 		for(int i = 0; i < annotations.length; i++) {
 			annotations[i] = new Annotation(raf);
 		}
